% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_by_tag.R
\name{boxplot_by_tag}
\alias{boxplot_by_tag}
\title{boxplot_by_tag}
\usage{
boxplot_by_tag(df, resp = "tp_med", tag)
}
\arguments{
\item{df}{a dataframe of bundled procedures}

\item{resp}{a price prediction, the default is tp_med}

\item{tag}{a tag of interest which we want to see side by side distributions for}
}
\value{
ggplot object
}
\description{
boxplot_by_tag
}
