% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btbv3.R
\name{btbv3}
\alias{btbv3}
\title{btbv3}
\usage{
btbv3(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
groups each bundle by a doctor facility pair based upon cleaned clean npi service and bill provider names.
}
