% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btbv5_medi.R
\name{btbv5_medi}
\alias{btbv5_medi}
\title{btbv5_medi}
\usage{
btbv5_medi(df, osa_group_p, osa_class_p, osa_specialization_p)
}
\arguments{
\item{osa_specialization_p}{}
}
\value{

}
\description{
btbv5_medi
}
