% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btbv5_radi.R
\name{btbv5_radi}
\alias{btbv5_radi}
\title{btbv5_radi}
\usage{
btbv5_radi(df, osa_group_p, osa_class_p, osa_specialization_p)
}
\arguments{
\item{osa_specialization_p}{}
}
\value{

}
\description{
btbv5_radi
}
