% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btbv5_radi_fac_only.R
\name{btbv5_radi_fac_only}
\alias{btbv5_radi_fac_only}
\title{btbv5_radi_fac_only}
\usage{
btbv5_radi_fac_only(df, osa_group_p, osa_class_p, osa_specialization_p)
}
\arguments{
\item{osa_specialization_p}{}
}
\value{

}
\description{
btbv5_radi_fac_only
}
