% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_primary_doctor.R
\name{calculate_primary_doctor}
\alias{calculate_primary_doctor}
\title{calculate_primary_doctor}
\usage{
calculate_primary_doctor(
  doctor_npi1,
  doctor_npi2,
  group_reqs = NA,
  class_reqs = NA,
  specialization_reqs = NA
)
}
\arguments{
\item{specialization_reqs}{}
}
\value{

}
\description{
calculate_primary_doctor
}
