% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_number_provider.R
\name{count_number_provider}
\alias{count_number_provider}
\title{count_number_provider}
\usage{
count_number_provider(doctor_str)
}
\arguments{
\item{doctor_str}{}
}
\value{

}
\description{
count_number_provider
}
