% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clean_sp_bp_names.R
\name{create_clean_sp_bp_names}
\alias{create_clean_sp_bp_names}
\title{create_clean_sp_bp_names}
\usage{
create_clean_sp_bp_names(
  surg_sp_name_clean,
  surg_bp_name_clean,
  faci_sp_name_clean,
  faci_bp_name_clean
)
}
\arguments{
\item{surg_sp_name_clean}{surgery service provider npi name}

\item{surg_bp_name_clean}{surgery bill provider npi name}

\item{faci_sp_name_clean}{facility service provider npi name}

\item{faci_bp_name_clean}{facility bill provider npi name}
}
\value{
list( doctor_str,facility_str)
}
\description{
create_clean_sp_bp_names
}
