% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clean_sp_bp_names_no_anes.R
\name{create_clean_sp_bp_names_no_anes}
\alias{create_clean_sp_bp_names_no_anes}
\title{create_clean_sp_bp_names_no_anes}
\usage{
create_clean_sp_bp_names_no_anes(
  surg_pc_codes,
  surg_sp_name_clean,
  surg_bp_name_clean,
  faci_sp_name_clean,
  faci_bp_name_clean
)
}
\arguments{
\item{faci_bp_name_clean}{}
}
\value{

}
\description{
create_clean_sp_bp_names_no_anes
}
