% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_out_most_important_facility.R
\name{filter_out_most_important_facility}
\alias{filter_out_most_important_facility}
\title{filter_out_most_important_facility}
\usage{
filter_out_most_important_facility(facility_str)
}
\arguments{
\item{facility_str}{}
}
\value{

}
\description{
filter_out_most_important_facility
}
