% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_strange_hosp_combo.R
\name{filter_strange_hosp_combo}
\alias{filter_strange_hosp_combo}
\title{filter_strange_hosp_combo}
\usage{
filter_strange_hosp_combo(facility_str)
}
\arguments{
\item{facility_str}{}
}
\value{

}
\description{
filter_strange_hosp_combo
}
