% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extra_doctors.R
\name{find_extra_doctors}
\alias{find_extra_doctors}
\title{find_extra_doctors}
\usage{
find_extra_doctors(doctor_str)
}
\arguments{
\item{doctor_str}{a string of cleaned doctor names}
}
\value{
the number of doctors in the string
}
\description{
find_extra_doctors
}
