% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_important_facility_npi.R
\name{get_most_important_facility_npi}
\alias{get_most_important_facility_npi}
\title{Title}
\usage{
get_most_important_facility_npi(
  most_important_fac,
  facility_str,
  facility_npi_str
)
}
\arguments{
\item{facility_npi_str}{}
}
\value{

}
\description{
Title
}
