% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_npi_standard_name.R
\name{get_npi_standard_name}
\alias{get_npi_standard_name}
\title{get_npi_standard_name}
\usage{
get_npi_standard_name(npi_bundle)
}
\arguments{
\item{npi_bundle}{}
}
\value{

}
\description{
get_npi_standard_name
}
