% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tag_density_information.R
\name{get_tag_density_information}
\alias{get_tag_density_information}
\title{get_tag_density_information}
\usage{
get_tag_density_information(df, tag)
}
\arguments{
\item{df}{a dataframe of bundled procedures}

\item{tag}{a tag of interest}
}
\value{
a list of ggobjects (a grid of density distributions and boxplots), and a stats table for the splitting on that paramter.
}
\description{
get_tag_density_information
}
\examples{
cscopy \%>\% get_tag_density_information(tag="surg_bun_t_biopsy")
}
