% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in_nerve_blocks.R
\name{not_in_nerve_blocks}
\alias{not_in_nerve_blocks}
\title{not_in_nerve_blocks}
\usage{
not_in_nerve_blocks(pc_str)
}
\arguments{
\item{pc_str}{}
}
\value{

}
\description{
checks if the pc_str is in a nerve block range.
}
