% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_med_density.R
\name{plot_med_density}
\alias{plot_med_density}
\title{plot_med_density}
\usage{
plot_med_density(
  df,
  temp_title = "Density Plot",
  temp_ylab = "Count",
  temp_xlab = "Median Bundled Procedure Price"
)
}
\arguments{
\item{df}{dataframe of bundled procedures}

\item{temp_title}{the title of the plot}

\item{temp_ylab}{the title of the y axis}

\item{temp_xlab}{the title of the x axis}
}
\value{

}
\description{
plot_med_density
}
