% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_price_vs_duration.R
\name{plot_price_vs_duration}
\alias{plot_price_vs_duration}
\title{plot_price_vs_duration
a ggplot object which has the tp_med vs the mean duration of the bundled procedures}
\usage{
plot_price_vs_duration(df)
}
\arguments{
\item{df}{a dataframe of bundled procedures}
}
\value{
a ggplot object
}
\description{
plot_price_vs_duration
a ggplot object which has the tp_med vs the mean duration of the bundled procedures
}
\examples{
cscopy \%>\% plot_pric_vs_duration()
}
