% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridge_tag.R
\name{plot_ridge_tag}
\alias{plot_ridge_tag}
\title{plot_ridge_tag}
\usage{
plot_ridge_tag(df, tag, resp = "tp_med")
}
\arguments{
\item{df}{a dataframe of bundled procedures}

\item{tag}{a tag of interest}

\item{resp}{the response that we want to compare the tag against. the default is tp_med}
}
\value{

}
\description{
plot_ridge_tag
}
\examples{
#cscopy \%>\% plot_ridge_tag(tag="surg_bun_t_biopsy")
}
